Adapta el backend al stock por sucursal,transalado y movimientos

+++ src/main/java/com/sistemapos/sistematextil/util/producto/ProductoVarianteResumenRow.java
@@ -17,7 +17,7 @@ public record ProductoVarianteResumenRow(
         Double precioOferta,
         LocalDateTime ofertaInicio,
         LocalDateTime ofertaFin,
-        Integer stock,
+        Long stock,
         String estado
 ) {
 }
